+++ search/bestfs.py
@@ -36,12 +36,12 @@ def best_first_search(start_node: Node):
     return None
 
 
-def stepped_best_first_search(node: Node, open: list[Node] | None, closed: set): 
+def stepped_best_first_search(node: Node, open: list[Node] | None, closed: set):
 
     if open is None:
         open = []
         heapq.heappush(open, node)
-    
+
     parent = heapq.heappop(open)
 
     if parent.board.goal_test():
@@ -55,5 +55,5 @@ def stepped_best_first_search(node: Node, open: list[Node] | None, closed: set):
 
         for child in children:
             heapq.heappush(open, Node(child, parent))
-        
+
         return False, parent.board, open, closed
